Fix import issues caused by moving browser test to root
