+++ browserTests/browserTest.js
@@ -3,7 +3,7 @@ import { Selector, ClientFunction } from 'testcafe';
 
 import { waitForReact, ReactSelector } from 'testcafe-react-selectors';
 
-import config from '../../config';
+import config from '../config';
 const { server } = config;
 
 // TODO: move these to the related view folders
